desktop: Refactor movie blitting into its own file
