+++ desktop/src/gui.rs
@@ -1,15 +1,16 @@
+mod movie;
+
 use crate::custom_event::RuffleEvent;
+use crate::gui::movie::MovieView;
 use anyhow::{anyhow, Result};
 use egui::*;
 use ruffle_render_wgpu::backend::request_adapter_and_device;
 use ruffle_render_wgpu::descriptors::Descriptors;
 use ruffle_render_wgpu::utils::{format_list, get_backend_names};
-use std::borrow::Cow;
 use std::path::Path;
 use std::rc::Rc;
 use std::sync::Arc;
 use std::time::{Duration, Instant};
-use wgpu::util::DeviceExt;
 use winit::event_loop::{EventLoop, EventLoopProxy};
 use winit::window::Window;
 
@@ -25,21 +26,8 @@ pub struct GuiController {
     repaint_after: Duration,
     surface: wgpu::Surface,
     surface_format: wgpu::TextureFormat,
-    blit_bind_group_layout: wgpu::BindGroupLayout,
-    blit_pipeline: wgpu::RenderPipeline,
-    blit_sampler: wgpu::Sampler,
-    blit_vertices: wgpu::Buffer,
+    movie_view: MovieView,
 }
-
-// x y u v
-const BLIT_VERTICES: [[f32; 4]; 6] = [
-    [-1.0, 1.0, 0.0, 0.0],  // tl
-    [1.0, 1.0, 1.0, 0.0],   // tr
-    [1.0, -1.0, 1.0, 1.0],  // br
-    [1.0, -1.0, 1.0, 1.0],  // br
-    [-1.0, -1.0, 0.0, 1.0], // bl
-    [-1.0, 1.0, 0.0, 0.0],  // tl
-];
 
 impl GuiController {
     pub fn new(
@@ -80,104 +68,7 @@ impl GuiController {
             .cloned()
             .expect("At least one format should be supported");
 
-        let module = descriptors
-            .device
-            .create_shader_module(wgpu::ShaderModuleDescriptor {
-                label: None,
-                source: wgpu::ShaderSource::Wgsl(Cow::Borrowed(include_str!("blit.wgsl"))),
-            });
-        let bind_group_layout =
-            descriptors
-                .device
-                .create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
-                    label: None,
-                    entries: &[
-                        wgpu::BindGroupLayoutEntry {
-                            binding: 0,
-                            visibility: wgpu::ShaderStages::FRAGMENT,
-                            ty: wgpu::BindingType::Texture {
-                                multisampled: false,
-                                sample_type: wgpu::TextureSampleType::Float { filterable: true },
-                                view_dimension: wgpu::TextureViewDimension::D2,
-                            },
-                            count: None,
-                        },
-                        wgpu::BindGroupLayoutEntry {
-                            binding: 1,
-                            visibility: wgpu::ShaderStages::FRAGMENT,
-                            ty: wgpu::BindingType::Sampler(wgpu::SamplerBindingType::Filtering),
-                            count: None,
-                        },
-                    ],
-                });
-        let sampler = descriptors.device.create_sampler(&wgpu::SamplerDescriptor {
-            mag_filter: wgpu::FilterMode::Linear,
-            min_filter: wgpu::FilterMode::Linear,
-            ..Default::default()
-        });
-        let pipeline_layout =
-            descriptors
-                .device
-                .create_pipeline_layout(&wgpu::PipelineLayoutDescriptor {
-                    label: None,
-                    bind_group_layouts: &[&bind_group_layout],
-                    push_constant_ranges: &[],
-                });
-        let pipeline = descriptors
-            .device
-            .create_render_pipeline(&wgpu::RenderPipelineDescriptor {
-                label: None,
-                layout: Some(&pipeline_layout),
-                vertex: wgpu::VertexState {
-                    entry_point: "vs_main",
-                    module: &module,
-                    buffers: &[wgpu::VertexBufferLayout {
-                        array_stride: 4 * 4,
-                        step_mode: wgpu::VertexStepMode::Vertex,
-                        // 0: vec2 position
-                        // 1: vec2 texture coordinates
-                        attributes: &wgpu::vertex_attr_array![0 => Float32x2, 1 => Float32x2],
-                    }],
-                },
-                primitive: wgpu::PrimitiveState {
-                    topology: wgpu::PrimitiveTopology::TriangleList,
-                    unclipped_depth: false,
-                    conservative: false,
-                    cull_mode: None,
-                    front_face: wgpu::FrontFace::default(),
-                    polygon_mode: wgpu::PolygonMode::default(),
-                    strip_index_format: None,
-                },
-                depth_stencil: None,
-                multisample: wgpu::MultisampleState {
-                    alpha_to_coverage_enabled: false,
-                    count: 1,
-                    mask: !0,
-                },
-
-                fragment: Some(wgpu::FragmentState {
-                    module: &module,
-                    entry_point: if surface_format.is_srgb() {
-                        "fs_main_srgb_framebuffer"
-                    } else {
-                        "fs_main_linear_framebuffer"
-                    },
-                    targets: &[Some(wgpu::ColorTargetState {
-                        format: surface_format,
-                        blend: Some(wgpu::BlendState::REPLACE),
-                        write_mask: wgpu::ColorWrites::ALL,
-                    })],
-                }),
-                multiview: None,
-            });
-        let vertices = descriptors
-            .device
-            .create_buffer_init(&wgpu::util::BufferInitDescriptor {
-                label: None,
-                contents: bytemuck::cast_slice(&BLIT_VERTICES),
-                usage: wgpu::BufferUsages::VERTEX,
-            });
-
+        let game_view = MovieView::new(&descriptors.device, surface_format);
         let egui_renderer = egui_wgpu::Renderer::new(&descriptors.device, surface_format, None, 1);
         let event_loop = event_loop.create_proxy();
         let gui = RuffleGui::new(event_loop);
@@ -192,10 +83,7 @@ impl GuiController {
             repaint_after: Duration::ZERO,
             surface,
             surface_format,
-            blit_bind_group_layout: bind_group_layout,
-            blit_pipeline: pipeline,
-            blit_sampler: sampler,
-            blit_vertices: vertices,
+            movie_view: game_view,
         })
     }
 
@@ -231,7 +119,6 @@ impl GuiController {
             .surface
             .get_current_texture()
             .expect("Surface became unavailable");
-        let movie_view = &movie.create_view(&Default::default());
 
         let raw_input = self.egui_winit.take_egui_input(&self.window);
         let full_output = self.egui_ctx.run(raw_input, |context| {
@@ -279,40 +166,24 @@ impl GuiController {
 
         {
             let surface_view = surface_texture.texture.create_view(&Default::default());
-            let blit_bind_group =
-                self.descriptors
-                    .device
-                    .create_bind_group(&wgpu::BindGroupDescriptor {
-                        label: None,
-                        layout: &self.blit_bind_group_layout,
-                        entries: &[
-                            wgpu::BindGroupEntry {
-                                binding: 0,
-                                resource: wgpu::BindingResource::TextureView(movie_view),
-                            },
-                            wgpu::BindGroupEntry {
-                                binding: 1,
-                                resource: wgpu::BindingResource::Sampler(&self.blit_sampler),
-                            },
-                        ],
-                    });
+
+            // First draw the movie - this also clears the surface
+            self.movie_view
+                .render(&self.descriptors.device, &mut encoder, movie, &surface_view);
+
+            // Then any UI
             let mut render_pass = encoder.begin_render_pass(&wgpu::RenderPassDescriptor {
                 color_attachments: &[Some(wgpu::RenderPassColorAttachment {
                     view: &surface_view,
                     resolve_target: None,
                     ops: wgpu::Operations {
-                        load: wgpu::LoadOp::Clear(wgpu::Color::WHITE),
+                        load: wgpu::LoadOp::Load,
                         store: true,
                     },
                 })],
                 depth_stencil_attachment: None,
                 label: Some("egui_render"),
             });
-
-            render_pass.set_pipeline(&self.blit_pipeline);
-            render_pass.set_bind_group(0, &blit_bind_group, &[]);
-            render_pass.set_vertex_buffer(0, self.blit_vertices.slice(..));
-            render_pass.draw(0..6, 0..1);
 
             self.egui_renderer
                 .render(&mut render_pass, &clipped_primitives, &screen_descriptor);
